Removed LandmarkAnnotation class as GNLandmarks are now annotations themselves.
Be smarter about annotation removal and adding: only remove annotations
that are actually gone and only add actual new annotations. Also, be
smarter about computing the region to set the map to. TODO: figure out
a way to let the user pan around the map without it reseting them on
every backend refresh.